Renamed utility module to "validation." Added docstrings to all validation functions
